Refactor DjangocmsVueJsMixin to allow for static url generation
Since we don't want to print hundreds of detail model urls in the frontend, we can generate routes à la
en/api/model/:model in the menu_helpers of a model. The current implementation doesn't allow
this and needs some nasty hacks sind this needs to be done with static methods. This refactor
introduces said methods.