Fix formatting, deprecate vue_js_link on every listview
get_frontend_list_data_dict is always called by a VueRouterListView
and automatically appended the pk and a detail route to the generated
json. Since this was hard coded and sometimes labels/attributes change,
it was deprecated in this commit. If you still need it feel free to
overload the get_frontend_list_data_dict and call it manually.